chore(cca): Refactor, break up into files
